# improve settings cascade performance